+++ packages/net-kit/lib/src/manager/params/net_kit_params.dart
@@ -12,6 +12,8 @@ class NetKitParams {
     required this.testMode,
     required this.loggerEnabled,
     required this.logLevel,
+    required this.accessTokenKey,
+    required this.refreshTokenKey,
     this.interceptor,
     this.internetStatusSubscription,
   });
@@ -40,6 +42,18 @@ class NetKitParams {
 
   /// The log level for the logger
   final LogLevel logLevel;
+
+  /// The access token key.
+  /// The default value is ['Authorization']
+  /// The access token key is used to get the access token from the headers
+  /// of the network responses
+  final String accessTokenKey;
+
+  /// The refresh token key.
+  /// The default value is ['Refresh-Token']
+  /// The refresh token key is used to get the refresh token from the headers
+  /// of the network responses
+  final String refreshTokenKey;
 
   /// Whether the NetKit logger is enabled
   bool get isNetKitLoggerEnabled => logLevel != LogLevel.off;