+++ packages/net-kit/lib/src/utility/converter.dart
@@ -2,16 +2,13 @@ import 'dart:core';
 
 import '../enum/http_status_codes.dart';
 import '../manager/error/api_exception.dart';
-import '../manager/logger/i_net_kit_logger.dart';
 import '../model/i_net_kit_model.dart';
 import 'typedef/request_type_def.dart';
 
 /// The Converters class contains methods to convert data
 class Converter {
   /// The constructor for the Converters class
-  Converter({required INetKitLogger logger}) : _logger = logger;
-
-  final INetKitLogger _logger;
+  const Converter();
 
   /// Converts the data to a list of models
   /// It takes in the following parameters:
@@ -23,18 +20,11 @@ class Converter {
   List<T> toListModel<T extends INetKitModel>({
     required dynamic data,
     required T parseModel,
-    bool loggerEnabled = false,
   }) {
     try {
       /// If the data is a list, convert it to a list of models
       if (data is List<dynamic>) {
-        /// The stopwatch to measure the time taken to convert the data
-        Stopwatch? stopwatch;
-
-        /// Start the stopwatch if the logger is enabled
-        if (loggerEnabled) stopwatch = Stopwatch()..start();
-
-        final parsedList = data
+        return data
 
             /// Filter the data to only include maps
             .whereType<MapType>()
@@ -47,15 +37,6 @@ class Converter {
 
             /// Convert the converted models to a list
             .toList();
-
-        /// Log the event if the logger is enabled
-        if (loggerEnabled) {
-          _logListParsingEvent(
-            stopwatch: stopwatch,
-            length: parsedList.length,
-          );
-        }
-        return parsedList;
       }
 
       /// If the data is not a list, throw an exception
@@ -90,16 +71,5 @@ class Converter {
         statusCode: HttpStatuses.expectationFailed.code,
       );
     }
-  }
-
-  Future<void> _logListParsingEvent({
-    required int length,
-    Stopwatch? stopwatch,
-  }) async {
-    stopwatch?.stop();
-    _logger.info(
-      'Converter.toListModel,'
-      ' $length items: ${stopwatch?.elapsedMilliseconds}ms',
-    );
   }
 }