+++ packages/net-kit/test/mocks/mock_error_interceptor_handler.dart
@@ -1,0 +1,5 @@
+import 'package:dio/dio.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockErrorInterceptorHandler extends Mock
+    implements ErrorInterceptorHandler {}