+++ packages/net-kit/test/mocks/mock_net_kit_manager.dart
@@ -1,0 +1,6 @@
+import 'package:mocktail/mocktail.dart';
+import 'package:net_kit/net_kit.dart';
+
+class MockNetKitManager extends Mock implements NetKitManager {
+  MockNetKitManager();
+}